+++ domain/model/src/main/java/co/com/crediya/model/user/exceptions/SalaryOutOfRangeException.java
@@ -1,6 +1,6 @@
 package co.com.crediya.model.user.exceptions;
 
-import co.com.crediya.model.common.gateways.exceptions.DomainException;
+import co.com.crediya.model.common.exceptions.DomainException;
 
 import java.math.BigDecimal;
