+++ domain/usecase/src/main/java/co/com/crediya/exception/IdentityDocumentAlreadyExists.java
@@ -1,6 +1,6 @@
 package co.com.crediya.exception;
 
-import co.com.crediya.model.common.gateways.exceptions.DomainException;
+import co.com.crediya.model.common.exceptions.DomainException;
 
 public class IdentityDocumentAlreadyExists extends DomainException {
     public IdentityDocumentAlreadyExists(String identityDocument) {
