+++ domain/usecase/src/main/java/co/com/crediya/exception/UserNotFoundException.java
@@ -1,7 +1,5 @@
 package co.com.crediya.exception;
 
-import co.com.crediya.model.common.gateways.exceptions.DomainException;
-
 public class UserNotFoundException extends NotFoundException {
     public UserNotFoundException() {
         super("USER_NOT_FOUND", "Usuario no encontrado");
